% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{build_suggesters}
\alias{build_suggesters}
\title{Indexes the search suggestions}
\usage{
build_suggesters(DomainName)
}
\arguments{
\item{DomainName}{}
}
\description{
Indexes the search suggestions. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html#configuring-suggesters}{Configuring Suggesters} in the \emph{Amazon CloudSearch Developer Guide}.
}
\examples{

}
