% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{delete_analysis_scheme}
\alias{delete_analysis_scheme}
\title{Deletes an analysis scheme}
\usage{
delete_analysis_scheme(DomainName, AnalysisSchemeName)
}
\arguments{
\item{DomainName}{}

\item{AnalysisSchemeName}{The name of the analysis scheme you want to delete.}
}
\description{
Deletes an analysis scheme. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis-schemes.html}{Configuring Analysis Schemes} in the \emph{Amazon CloudSearch Developer Guide}.
}
\examples{

}
