% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{delete_domain}
\alias{delete_domain}
\title{Permanently deletes a search domain and all of its data}
\usage{
delete_domain(
  DomainName = "string"
)
}
\arguments{
\item{DomainName}{[required] The name of the domain you want to permanently delete.}
}
\description{
Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/deleting-domains.html}{Deleting a Search Domain} in the \emph{Amazon CloudSearch Developer Guide}.
}
\examples{

}
