% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{delete_expression}
\alias{delete_expression}
\title{Removes an \code{Expression} from the search domain}
\usage{
delete_expression(DomainName, ExpressionName)
}
\arguments{
\item{DomainName}{}

\item{ExpressionName}{The name of the \code{Expression} to delete.}
}
\description{
Removes an \code{Expression} from the search domain. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-expressions.html}{Configuring Expressions} in the \emph{Amazon CloudSearch Developer Guide}.
}
\examples{

}
