% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{delete_index_field}
\alias{delete_index_field}
\title{Removes an IndexField from the search domain}
\usage{
delete_index_field(DomainName, IndexFieldName)
}
\arguments{
\item{DomainName}{[required]}

\item{IndexFieldName}{[required] The name of the index field your want to remove from the domain's indexing options.}
}
\description{
Removes an \code{IndexField} from the search domain. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index-fields.html}{Configuring Index Fields} in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{delete_index_field(
  DomainName = "string",
  IndexFieldName = "string"
)
}
}

