% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{delete_suggester}
\alias{delete_suggester}
\title{Deletes a suggester}
\usage{
delete_suggester(DomainName, SuggesterName)
}
\arguments{
\item{DomainName}{[required]}

\item{SuggesterName}{[required] Specifies the name of the suggester you want to delete.}
}
\description{
Deletes a suggester. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html}{Getting Search Suggestions} in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{delete_suggester(
  DomainName = "string",
  SuggesterName = "string"
)
}
}

