% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{describe_scaling_parameters}
\alias{describe_scaling_parameters}
\title{Gets the scaling parameters configured for a domain}
\usage{
describe_scaling_parameters(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see \href{http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-scaling-options.html}{Configuring Scaling Options} in the \emph{Amazon CloudSearch Developer Guide}.
}
\section{Accepted Parameters}{
\preformatted{describe_scaling_parameters(
  DomainName = "string"
)
}
}

