% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{index_documents}
\alias{index_documents}
\title{Tells the search domain to start indexing its documents using the latest indexing options}
\usage{
index_documents(DomainName)
}
\arguments{
\item{DomainName}{[required]}
}
\description{
Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is \code{RequiresIndexDocuments}.
}
\section{Accepted Parameters}{
\preformatted{index_documents(
  DomainName = "string"
)
}
}

