% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_operations.R
\name{list_domain_names}
\alias{list_domain_names}
\title{Lists all search domains owned by an account}
\usage{
list_domain_names()
}
\description{
Lists all search domains owned by an account.
}
\examples{

}
