% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudsearch_package.R
\docType{package}
\name{paws.cloudsearch-package}
\alias{paws.cloudsearch}
\alias{paws.cloudsearch-package}
\title{paws.cloudsearch: Amazon CloudSearch}
\description{
You use the Amazon CloudSearch configuration
    service to create, configure, and manage search domains. Configuration
    service requests are submitted using the AWS Query protocol. AWS Query
    requests are HTTP or HTTPS requests submitted via HTTP GET or POST
    with a query parameter named Action.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
